Fix Moktang Circulars
