+++ src/mahoji/lib/abstracted_commands/moktangCommand.ts
@@ -1,27 +1,17 @@
-import { spoiler, userMention } from '@discordjs/builders';
-import { randInt, Time } from 'e';
+import { spoiler } from '@discordjs/builders';
+import { Time } from 'e';
 import { Bank } from 'oldschooljs';
 
-import { Events } from '../../../lib/constants';
 import { dwarvenOutfit } from '../../../lib/data/CollectionsExport';
-import { isDoubleLootActive } from '../../../lib/doubleLoot';
 import { trackLoot } from '../../../lib/lootTrack';
-import { MOKTANG_ID, MoktangLootTable } from '../../../lib/minions/data/killableMonsters/custom/bosses/Moktang';
 import { SkillsEnum } from '../../../lib/skilling/types';
 import { PercentCounter } from '../../../lib/structures/PercentCounter';
-import { ActivityTaskOptions } from '../../../lib/types/minions';
+import { MoktangTaskOptions } from '../../../lib/types/minions';
 import { formatDuration, itemNameFromID } from '../../../lib/util';
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { calcMaxTripLength } from '../../../lib/util/calcMaxTripLength';
-import { formatOrdinal } from '../../../lib/util/formatOrdinal';
-import { handleTripFinish } from '../../../lib/util/handleTripFinish';
-import { makeBankImage } from '../../../lib/util/makeBankImage';
 import resolveItems from '../../../lib/util/resolveItems';
 import { updateBankSetting } from '../../../lib/util/updateBankSetting';
-
-export interface MoktangTaskOptions extends ActivityTaskOptions {
-	qty: number;
-}
 
 const requiredPickaxes = resolveItems(['Crystal pickaxe', 'Volcanic pickaxe', 'Dwarven pickaxe', 'Dragon pickaxe']);
 
@@ -93,72 +83,4 @@ export async function moktangCommand(user: MUser, channelID: string, inputQuanti
 **Boosts:** ${timeToKill.messages.join(', ')} ${
 		timeToKill.missed.length > 0 ? spoiler(timeToKill.missed.join(', ')) : ''
 	}`;
-}
-
-export async function moktangActivity(data: MoktangTaskOptions) {
-	const { userID, qty } = data;
-	const user = await mUserFetch(userID);
-
-	await user.incrementKC(MOKTANG_ID, qty);
-
-	let loot = new Bank();
-
-	for (let i = 0; i < qty; i++) {
-		loot.add(MoktangLootTable.roll());
-	}
-	if (isDoubleLootActive(data.duration)) {
-		loot.multiply(2);
-		data.cantBeDoubled = true;
-	}
-
-	const res = await user.addItemsToBank({ items: loot, collectionLog: true });
-	await updateBankSetting('moktang_loot', loot);
-	await trackLoot({
-		duration: data.duration,
-		totalLoot: loot,
-		type: 'Monster',
-		changeType: 'loot',
-		id: 'moktang',
-		kc: qty,
-		users: [
-			{
-				id: user.id,
-				loot,
-				duration: data.duration
-			}
-		]
-	});
-
-	const xpStr = await user.addXP({
-		skillName: SkillsEnum.Mining,
-		amount: user.skillLevel(SkillsEnum.Mining) * 2000 * qty,
-		duration: data.duration,
-		multiplier: false,
-		masterCapeBoost: true
-	});
-
-	const image = await makeBankImage({
-		bank: res.itemsAdded,
-		title: `Loot From ${qty} Moktang`,
-		user,
-		previousCL: res.previousCL
-	});
-
-	const newKC = user.getKC(MOKTANG_ID);
-	for (const item of resolveItems(['Igne gear frame', 'Mini moktang'])) {
-		if (loot.has(item)) {
-			globalClient.emit(
-				Events.ServerNotification,
-				`**${user.usernameOrMention}'s** minion just received their ${formatOrdinal(
-					user.cl.amount(item)
-				)} ${itemNameFromID(item)} on their ${formatOrdinal(randInt(newKC - qty, newKC))} kill!`
-			);
-		}
-	}
-
-	let str = `${userMention(data.userID)}, ${user.minionName} finished killing ${qty}x Moktang. Received ${loot}.
-
-${xpStr}`;
-
-	handleTripFinish(user, data.channelID, str, image.file.attachment, data, loot);
 }
